Conversation between user and selected contact
This is the main chat area where messages are displayed. 
A messages array (received from parent as prop) containing only the conversation between the selected contact and you(user) is mapped to display conversation between user and selected contact.
Message sent by user are displayed on right and received are displayed on left (like WhatsApp).