Component to manage Chats (messages)
Component to send and view messages. 
Contains main logic to send and display messages which is passed to child as props.
Contains message list which is passed to child component (ChatArea) as a prop to display messages. Only those messages are passed to the child to display, which corresponds to the conversation between user and the currently selected contact for conversation.
If no contact is selected (e.g. initially when app loads), a welcome page is displayed, with details of the user like name and description. (conditional rendering)