Main App component
Contains user's Contacts list. And maintains the state of currently selected contact for conversation.
Initially when no contact is selected on app load, the User itself is selected and his own info like name and description are shown as Welcome Page, instead of chat area.
Chat Area is only displayed when a contact is selected for conversation.